FvwmPager: Splitup large source files.
Split up the two monolithic source FvwmPager.c and x_pager.c source
files into smaller files to help manage the code. The new structure is.

FvwmPager.c  - main event loop
fpmonitors.c - fpmonitor methods.
init_pager.c - collection of all the initialization methods.
messages.c   - respond to messages from fvwm.
x_pager.c    - main still fairly large handling of the X windows
               that build the pager, this includes processing events
               and computing the position of all the windows.
x_update.c   - update_foo methods for x_pager.c to update locations
               and window graphics.
