+++ modules/FvwmPager/x_pager.c
@@ -30,55 +30,46 @@
 #include <signal.h>
 #include <stdbool.h>
 
-#include "libs/FScreen.h"
-#include "libs/fvwmlib.h"
 #include "libs/FShape.h"
 #include "libs/Module.h"
-#include "libs/Colorset.h"
-#include "libs/Graphics.h"
-#include "fvwm/fvwm.h"
-#include "libs/PictureGraphics.h"
 #include "libs/FEvent.h"
+#include "fvwm/fvwm.h"
 #include "FvwmPager.h"
 
-static Atom wm_del_win;
-static int desk_w = 0;
-static int desk_h = 0;
-static int label_h = 0;
+Atom wm_del_win;
+int desk_w = 0;
+int desk_h = 0;
+int label_h = 0;
 static int Wait = 0;
 static int MyVx, MyVy;		/* copy of Scr.Vx/y for drag logic */
 static struct fpmonitor *ScrollFp = NULL;	/* Stash monitor drag logic */
 
-/* assorted gray bitmaps for decorative borders */
-#define g_width 2
-#define g_height 2
-static char g_bits[] = {0x02, 0x01};
-
-#define l_g_width 4
-#define l_g_height 2
-static char l_g_bits[] = {0x08, 0x02};
-
-#define s_g_width 4
-#define s_g_height 4
-static char s_g_bits[] = {0x01, 0x02, 0x04, 0x08};
-
-#define label_border g_width
-
-static int FvwmErrorHandler(Display *, XErrorEvent *);
+static void do_scroll(
+	int sx, int sy, bool do_send_message, bool is_message_recieved);
+static int _pred_weed_window_events(
+	Display *display, XEvent *current_event, XPointer arg);
+static void discard_events(long event_type, Window w, XEvent *last_ev);
 static rectangle CalcGeom(PagerWindow *, bool);
-static void HideWindow(PagerWindow *, Window);
-static void MoveResizeWindow(PagerWindow *, bool, bool);
-static void DrawInBalloonWindow(void);
-static rectangle set_vp_size_and_loc(struct fpmonitor *, bool is_icon);
-static struct fpmonitor *fpmonitor_from_xy(int x, int y);
-static struct fpmonitor *fpmonitor_from_n(int n);
-static int fpmonitor_count(void);
 static void fvwmrec_to_pager(rectangle *, bool, struct fpmonitor *);
 static void pagerrec_to_fvwm(rectangle *, bool, struct fpmonitor *);
-static char *get_label(const PagerWindow *pw,const char *fmt);
-static int is_window_visible(PagerWindow *t);
+static void UpdateWindowShape(void);
+static void HandleEnterNotify(XEvent *Event);
+static void HandleExpose(XEvent *Event);
 static void draw_desk_label(const char *label, const char *small, int desk,
 			int x, int y, int width, int height, bool hilight);
+static void draw_icon_grid(int erase);
+static void SwitchToDesk(int Desk, struct fpmonitor *m);
+static void SwitchToDeskAndPage(int Desk, XEvent *Event);
+static void IconSwitchPage(XEvent *Event);
+static void HideWindow(PagerWindow *, Window);
+static void MoveResizeWindow(PagerWindow *, bool, bool);
+static int is_window_visible(PagerWindow *t);
+static void Scroll(int x, int y, int Desk, bool do_scroll_icon);
+static void MoveWindow(XEvent *Event);
+static void IconMoveWindow(XEvent *Event, PagerWindow *t);
+static void setup_balloon_window(PagerWindow *t);
+static void InsertExpand(char **dest, char *s);
+static void DrawInBalloonWindow(void);
 
 #define  MAX_UNPROCESSED_MESSAGES 1
 /* sums up pixels to scroll. If do_send_message is true a Scroll command is
@@ -166,83 +157,6 @@ static void discard_events(long event_type, Window w, XEvent *last_ev)
 	return;
 }
 
-int fpmonitor_get_all_widths(void)
-{
-	struct fpmonitor *fp = TAILQ_FIRST(&fp_monitor_q);
-
-	if (fp->scr_width <= 0)
-		return (monitor_get_all_widths());
-
-	return (fp->scr_width);
-}
-
-int fpmonitor_get_all_heights(void)
-{
-	struct fpmonitor *fp = TAILQ_FIRST(&fp_monitor_q);
-
-	if (fp->scr_height <= 0)
-		return (monitor_get_all_heights());
-	return (fp->scr_height);
-}
-
-int fpmonitor_count(void)
-{
-	struct fpmonitor *fp;
-	int c = 0;
-
-	TAILQ_FOREACH(fp, &fp_monitor_q, entry)
-		if (!fp->disabled)
-			c++;
-	return (c);
-}
-
-static struct fpmonitor *fpmonitor_from_n(int n)
-{
-	struct fpmonitor *fp;
-	int c = 0;
-
-	TAILQ_FOREACH(fp, &fp_monitor_q, entry) {
-		if (fp->disabled)
-			continue;
-		if (c == n)
-			return fp;
-		c++;
-	}
-	return fp;
-}
-
-struct fpmonitor *fpmonitor_from_desk(int desk)
-{
-	struct fpmonitor *fp;
-
-	TAILQ_FOREACH(fp, &fp_monitor_q, entry)
-		if (!fp->disabled && fp->m->virtual_scr.CurrentDesk == desk)
-			return fp;
-
-	return NULL;
-}
-
-static struct fpmonitor *fpmonitor_from_xy(int x, int y)
-{
-	struct fpmonitor *fp;
-
-	if (monitor_to_track == NULL && monitor_mode != MONITOR_TRACKING_G) {
-		x %= fpmonitor_get_all_widths();
-		y %= fpmonitor_get_all_heights();
-
-		TAILQ_FOREACH(fp, &fp_monitor_q, entry)
-			if (x >= fp->m->si->x &&
-			    x < fp->m->si->x + fp->m->si->w &&
-			    y >= fp->m->si->y &&
-			    y < fp->m->si->y + fp->m->si->h)
-				return fp;
-	} else {
-		return fpmonitor_this(NULL);
-	}
-
-	return NULL;
-}
-
 static rectangle CalcGeom(PagerWindow *t, bool is_icon)
 {
 	/* Place initial rectangle off screen. */
@@ -311,306 +225,6 @@ static rectangle CalcGeom(PagerWindow *t, bool is_icon)
 		}
 	}
 	return rec;
-}
-
-/*
- *
- *  Procedure:
- *	Initialize_viz_pager - creates a temp window of the correct visual
- *	so that pixmaps may be created for use with the main window
- */
-void initialize_viz_pager(void)
-{
-	XSetWindowAttributes attr;
-	XGCValues xgcv;
-
-	/* FIXME: I think that we only need that Pdepth ==
-	 * DefaultDepth(dpy, Scr.screen) to use the Scr.Root for Scr.pager_w */
-	if (Pdefault)
-	{
-		Scr.pager_w = Scr.Root;
-	}
-	else
-	{
-		attr.background_pixmap = None;
-		attr.border_pixel = 0;
-		attr.colormap = Pcmap;
-		Scr.pager_w = XCreateWindow(
-			dpy, Scr.Root, -10, -10, 10, 10, 0, Pdepth,
-			InputOutput, Pvisual,
-			CWBackPixmap|CWBorderPixel|CWColormap, &attr);
-		Scr.NormalGC = fvwmlib_XCreateGC(dpy, Scr.pager_w, 0, &xgcv);
-	}
-	Scr.NormalGC = fvwmlib_XCreateGC(dpy, Scr.pager_w, 0, NULL);
-
-	xgcv.plane_mask = AllPlanes;
-	Scr.MiniIconGC = fvwmlib_XCreateGC(
-		dpy, Scr.pager_w, GCPlaneMask, &xgcv);
-
-	/* Transparent background are only allowed when the depth matched the
-	 * root */
-	if (Pdepth == DefaultDepth(dpy, Scr.screen))
-	{
-		default_pixmap = ParentRelative;
-	}
-
-	return;
-}
-
-void set_desk_size(bool update_label)
-{
-	if (update_label) {
-		label_h = 0;
-		if (use_desk_label)
-			label_h += Ffont->height + 2;
-		if (use_monitor_label)
-			label_h += Ffont->height + 2;
-	}
-
-	desk_w = (pwindow.width - Columns + 1) / Columns;
-	desk_h = (pwindow.height - Rows * label_h - Rows + 1) / Rows;
-
-	return;
-}
-
-void update_monitor_locations(int i)
-{
-	if (!HilightDesks)
-		return;
-
-	struct fpmonitor *fp;
-
-	TAILQ_FOREACH(fp, &fp_monitor_q, entry) {
-		if (fp->disabled)
-			continue;
-
-		rectangle vp = set_vp_size_and_loc(fp, false);
-
-		/* Conditions to show a monitor view port are:
-		 *   + Never show monitors not being tracked.
-		 *   + Always shown when CurrntDeskPerMonitor is in use.
-		 *   + Otherwise only show monitors on desk.
-		 */
-		if ((monitor_to_track == NULL || monitor_to_track == fp) &&
-		    ((CurrentDeskPerMonitor && fAlwaysCurrentDesk) ||
-		    fp->m->virtual_scr.CurrentDesk == i + desk1))
-		{
-			XMoveResizeWindow(dpy, fp->CPagerWin[i],
-				vp.x, vp.y, vp.width, vp.height);
-		} else {
-			XMoveResizeWindow(dpy, fp->CPagerWin[i],
-				-32768, -32768, vp.width, vp.height);
-		}
-	}
-}
-
-void update_monitor_backgrounds(int desk)
-{
-	if (!HilightDesks)
-		return;
-
-	rectangle vp;
-	struct fpmonitor *fp;
-	DeskStyle *style = Desks[desk].style;
-
-	TAILQ_FOREACH(fp, &fp_monitor_q, entry) {
-		if (CurrentDeskPerMonitor && fAlwaysCurrentDesk)
-			style = FindDeskStyle(fp->m->virtual_scr.CurrentDesk);
-
-		if (style->hi_cs < 0) {
-			if (style->hiPixmap != NULL)
-				XSetWindowBackgroundPixmap(dpy,
-					fp->CPagerWin[desk],
-					style->hiPixmap->picture);
-			else
-				XSetWindowBackground(dpy,
-					fp->CPagerWin[desk], style->hi_bg);
-		} else {
-			vp = set_vp_size_and_loc(fp, false);
-			SetWindowBackground(dpy,
-				fp->CPagerWin[desk], vp.width, vp.height,
-				&Colorset[style->hi_cs], Pdepth,
-				style->hi_bg_gc, True);
-		}
-		XLowerWindow(dpy, fp->CPagerWin[desk]);
-		XClearWindow(dpy, fp->CPagerWin[desk]);
-	}
-}
-
-void update_desk_background(int desk)
-{
-	DeskStyle *style = Desks[desk].style;
-
-	XSetWindowBorder(dpy, Desks[desk].title_w, style->fg);
-	XSetWindowBorder(dpy, Desks[desk].w, style->fg);
-
-	if (style->cs < 0) {
-		XSetWindowBorder(dpy, Desks[desk].title_w, style->fg);
-		XSetWindowBorder(dpy, Desks[desk].w, style->fg);
-		if (style->bgPixmap != NULL) {
-			if (style->use_label_pixmap)
-				XSetWindowBackgroundPixmap(dpy,
-					Desks[desk].title_w,
-					style->bgPixmap->picture);
-			else
-				XSetWindowBackgroundPixmap(dpy,
-					Desks[desk].w,
-					style->bgPixmap->picture);
-		} else {
-			XSetWindowBackground(dpy,
-				Desks[desk].title_w, style->bg);
-			XSetWindowBackground(dpy,
-				Desks[desk].w, style->bg);
-		}
-	} else {
-		if (style->use_label_pixmap)
-			SetWindowBackground(dpy,
-				Desks[desk].title_w, desk_w, desk_h + label_h,
-				&Colorset[style->cs], Pdepth, Scr.NormalGC,
-				True);
-		else
-			SetWindowBackground(dpy,
-				Desks[desk].w, desk_w, desk_h,
-				&Colorset[style->cs], Pdepth, Scr.NormalGC,
-				True);
-	}
-
-	XClearWindow(dpy, Desks[desk].title_w);
-	XClearWindow(dpy, Desks[desk].w);
-}
-
-void update_window_background(PagerWindow *t)
-{
-	if (t == NULL || t->PagerView == None || t->IconView == None)
-		return;
-
-	if (Pdepth < 2)
-	{
-		if (t == FocusWin) {
-			XSetWindowBackgroundPixmap(
-				dpy, t->PagerView, Scr.gray_pixmap);
-			XSetWindowBackgroundPixmap(
-				dpy, t->IconView, Scr.gray_pixmap);
-		} else if (IS_STICKY_ACROSS_DESKS(t) ||
-			   IS_STICKY_ACROSS_PAGES(t))
-		{
-			XSetWindowBackgroundPixmap(dpy, t->PagerView,
-						   Scr.sticky_gray_pixmap);
-			XSetWindowBackgroundPixmap(dpy, t->IconView,
-						   Scr.sticky_gray_pixmap);
-		} else {
-			XSetWindowBackgroundPixmap(dpy, t->PagerView,
-						   Scr.light_gray_pixmap);
-			XSetWindowBackgroundPixmap(dpy, t->IconView,
-						   Scr.light_gray_pixmap);
-		}
-		XClearArea(dpy, t->PagerView, 0, 0, 0, 0, True);
-		XClearArea(dpy, t->IconView, 0, 0, 0, 0, True);
-		return;
-	}
-
-	int cs = -1;
-	Pixel pix = None;
-	DeskStyle *style = FindDeskStyle(t->desk);
-
-	if (t == FocusWin) {
-		if (style->focus_cs >= 0) {
-			cs = style->focus_cs;
-		} else if (style->focus_bg) {
-			pix = style->focus_bg;
-		} else {
-			pix = Scr.focus_win_bg;
-		}
-	} else {
-		if (style->win_cs >= 0) {
-			cs = style->win_cs;
-		} else if (style->win_bg) {
-			pix = style->win_bg;
-		} else {
-			pix = t->back;
-		}
-	}
-
-	if (cs >= 0) {
-		SetWindowBackground(dpy, t->PagerView,
-			t->pager_view.width, t->pager_view.height,
-			&Colorset[cs], Pdepth, Scr.NormalGC, True);
-		SetWindowBackground(dpy, t->IconView,
-			t->icon_view.width, t->icon_view.height,
-			&Colorset[cs], Pdepth, Scr.NormalGC, True);
-	} else if (pix != None) {
-		XSetWindowBackground(dpy, t->PagerView, pix);
-		XSetWindowBackground(dpy, t->IconView, pix);
-	}
-
-	XClearArea(dpy, t->PagerView, 0, 0, 0, 0, True);
-	XClearArea(dpy, t->IconView, 0, 0, 0, 0, True);
-}
-
-/*
- *
- *  Procedure:
- *	Initialize_pager - creates the pager window, if needed
- *
- *  Inputs:
- *	x,y location of the window
- *
- */
-XSizeHints sizehints =
-{
-	(PWinGravity),		/* flags */
-	0, 0, 100, 100,		/* x, y, width and height (legacy) */
-	1, 1,			/* Min width and height */
-	0, 0,			/* Max width and height */
-	1, 1,			/* Width and height increments */
-	{0, 0}, {0, 0},		/* Aspect ratio */
-	1, 1,			/* base size */
-	(NorthWestGravity)	/* gravity */
-};
-
-/* create balloon window -- ric@giccs.georgetown.edu */
-void initialize_balloon_window(void)
-{
-	XGCValues xgcv;
-	unsigned long valuemask;
-	XSetWindowAttributes attributes;
-
-	Balloon.is_mapped = false;
-	if (!(Balloon.show_in_pager || Balloon.show_in_icon)) {
-		Scr.balloon_w = None;
-		return;
-	}
-
-	/* Set some defaults */
-	if (Balloon.border_width <= 0)
-		Balloon.border_width = DEFAULT_BALLOON_BORDER_WIDTH;
-	if (Balloon.y_offset == 0)
-		/* we don't allow y_offset of 0 because it allows direct
-		 * transit from pager window to balloon window, setting
-		 * up a LeaveNotify/EnterNotify event loop.
-		 */
-		Balloon.y_offset = DEFAULT_BALLOON_Y_OFFSET;
-	if (Balloon.label_format == NULL)
-		Balloon.label_format = fxstrdup("%i");
-
-	valuemask = CWOverrideRedirect | CWEventMask | CWColormap;
-	/* tell WM to ignore this window */
-	attributes.override_redirect = true;
-	attributes.event_mask = ExposureMask;
-	attributes.colormap = Pcmap;
-	/* now create the window */
-	Scr.balloon_w = XCreateWindow(
-		dpy, Scr.Root, 0, 0, /* coords set later */ 1, 1,
-		Balloon.border_width, Pdepth, InputOutput, Pvisual, valuemask,
-		&attributes);
-
-	/* Initialize, actual values will be updated later. */
-	valuemask = GCForeground;
-	xgcv.foreground = GetSimpleColor("black");
-	Balloon.Ffont = FlocaleLoadFont(dpy, BalloonFont, MyName);
-	Balloon.gc = fvwmlib_XCreateGC(
-		dpy, Scr.balloon_w, valuemask, &xgcv);
-	free(BalloonFont);
 }
 
 /* Computes pager size rectangle from fvwm size or visa versa */
@@ -687,8 +301,7 @@ pagerrec_to_fvwm(rectangle *rec, bool is_icon, struct fpmonitor *fp)
 	rec->y = (rec->y * m_height) / scale_h + offset_y;
 }
 
-static rectangle
-set_vp_size_and_loc(struct fpmonitor *m, bool is_icon)
+rectangle set_vp_size_and_loc(struct fpmonitor *m, bool is_icon)
 {
 	int Vx = 0, Vy = 0, vp_w, vp_h;
 	rectangle vp = {0, 0, 0, 0};
@@ -717,503 +330,6 @@ set_vp_size_and_loc(struct fpmonitor *m, bool is_icon)
 	vp.height = (Vy + vp_h) * scale_h / fp->virtual_scr.VHeight - vp.y;
 	return vp;
 }
-
-void initialise_common_pager_fragments(void)
-{
-	extern char *font_string, *smallFont;
-	DeskStyle *style = FindDeskStyle(-1);
-
-	/* I don't think that this is necessary - just let pager die */
-	/* domivogt (07-mar-1999): But it is! A window being moved in the pager
-	 * might die at any moment causing the Xlib calls to generate BadMatch
-	 * errors. Without an error handler the pager will die! */
-	XSetErrorHandler(FvwmErrorHandler);
-
-	wm_del_win = XInternAtom(dpy,"WM_DELETE_WINDOW",False);
-
-	/* load the font */
-	/* Note: "font" is always created, whether labels are used or not
-	   because a GC below is set to use a font. dje Dec 2001.
-	   OK, I fixed the GC below, but now something else is blowing up.
-	   Right now, I've got to do some Real Life stuff, so this kludge is
-	   in place, its still better than I found it.
-	   I hope that I've fixed this (olicha)
-	*/
-	Ffont = FlocaleLoadFont(dpy, font_string, MyName);
-
-	/* init our Flocale window string */
-	FlocaleAllocateWinString(&FwinString);
-
-	/* Check that shape extension exists. */
-	if (FHaveShapeExtension && ShapeLabels)
-	{
-		ShapeLabels = (FShapesSupported) ? 1 : 0;
-	}
-
-	if(smallFont != NULL)
-	{
-		FwindowFont = FlocaleLoadFont(dpy, smallFont, MyName);
-	}
-
-	/* Load pixmaps for mono use */
-	if (Pdepth < 2)
-	{
-		Scr.gray_pixmap = XCreatePixmapFromBitmapData(
-			dpy, Scr.pager_w, g_bits, g_width, g_height,
-			style->fg, style->bg, Pdepth);
-		Scr.light_gray_pixmap = XCreatePixmapFromBitmapData(
-			dpy, Scr.pager_w, l_g_bits, l_g_width, l_g_height,
-			style->fg, style->bg, Pdepth);
-		Scr.sticky_gray_pixmap = XCreatePixmapFromBitmapData(
-			dpy, Scr.pager_w, s_g_bits, s_g_width, s_g_height,
-			style->fg, style->bg, Pdepth);
-	}
-
-	/* Size the window */
-	if(Rows < 0)
-	{
-		if(Columns <= 0)
-		{
-			Columns = ndesks;
-			Rows = 1;
-		}
-		else
-	{
-			Rows = ndesks/Columns;
-			if(Rows*Columns < ndesks)
-			Rows++;
-		}
-	}
-	if(Columns < 0)
-	{
-		if (Rows == 0)
-			Rows = 1;
-		Columns = ndesks/Rows;
-		if(Rows*Columns < ndesks)
-			Columns++;
-	}
-
-	if(Rows*Columns < ndesks)
-	{
-		if (Columns == 0)
-			Columns = 1;
-		Rows = ndesks/Columns;
-		if (Rows*Columns < ndesks)
-			Rows++;
-	}
-	set_desk_size(true);
-}
-
-void initialize_fpmonitor_windows(struct fpmonitor *fp)
-{
-	/* No need to create windows if not using HilightDesks. */
-	if (!HilightDesks)
-		return;
-
-	int i;
-	unsigned long valuemask;
-	XSetWindowAttributes attributes;
-	rectangle vp = set_vp_size_and_loc(fp, false);
-
-	attributes.colormap = Pcmap;
-	for (i = 0; i < ndesks; i++) {
-		valuemask = CWColormap;
-		if (Desks[i].style->hiPixmap ||
-		    (Desks[i].style->hi_cs > -1 &&
-		    Colorset[Desks[i].style->hi_cs].pixmap))
-		{
-			valuemask |= CWBackPixmap;
-			attributes.background_pixmap = None; /* set later */
-		} else {
-			valuemask |= CWBackPixel;
-			attributes.background_pixel = Desks[i].style->hi_bg;
-		}
-
-		fp->CPagerWin[i] = XCreateWindow(dpy, Desks[i].w,
-				-32768, -32768, vp.width, vp.height, 0,
-				CopyFromParent, InputOutput,
-				CopyFromParent, valuemask,
-				&attributes);
-		XMapRaised(dpy, fp->CPagerWin[i]);
-		update_monitor_locations(i);
-		update_monitor_backgrounds(i);
-	}
-}
-
-void update_desk_style_gcs(DeskStyle *style)
-{
-	if (style->cs >= 0) {
-		style->fg = Colorset[style->cs].fg;
-		style->bg = Colorset[style->cs].bg;
-	}
-	if (style->hi_cs >= 0) {
-		style->hi_fg = Colorset[style->hi_cs].fg;
-		style->hi_bg = Colorset[style->hi_cs].bg;
-	}
-	if (style->win_cs >= 0) {
-		style->win_fg = Colorset[style->win_cs].fg;
-		style->win_bg = Colorset[style->win_cs].bg;
-	}
-	if (style->focus_cs >= 0) {
-		style->focus_fg = Colorset[style->focus_cs].fg;
-		style->focus_bg = Colorset[style->focus_cs].bg;
-	}
-	if (style->balloon_cs >= 0) {
-		style->balloon_fg = Colorset[style->balloon_cs].fg;
-		style->balloon_bg = Colorset[style->balloon_bg].bg;
-	}
-
-	XSetForeground(dpy, style->label_gc, style->fg);
-	XSetForeground(dpy, style->dashed_gc, style->fg);
-	XSetForeground(dpy, style->hi_bg_gc, style->hi_bg);
-	XSetForeground(dpy, style->hi_fg_gc, style->hi_fg);
-	if (WindowBorders3d && style->win_cs > 0) {
-		XSetForeground(dpy, style->win_hi_gc,
-			       Colorset[style->win_cs].hilite);
-		XSetForeground(dpy, style->win_sh_gc,
-			       Colorset[style->win_cs].shadow);
-	}
-	if (WindowBorders3d && style->focus_cs > 0) {
-		XSetForeground(dpy, style->focus_hi_gc,
-			       Colorset[style->focus_cs].hilite);
-		XSetForeground(dpy, style->focus_sh_gc,
-			       Colorset[style->focus_cs].shadow);
-	}
-}
-
-void initialize_desk_style_gcs(DeskStyle *style)
-{
-	XGCValues gcv;
-	char dash_list[2];
-
-	/* Desk labels GC. */
-	gcv.foreground = style->fg;
-	if (Ffont && Ffont->font) {
-		gcv.font = Ffont->font->fid;
-		style->label_gc = fvwmlib_XCreateGC(
-			dpy, Scr.pager_w, GCForeground | GCFont, &gcv);
-	} else {
-		style->label_gc = fvwmlib_XCreateGC(
-			dpy, Scr.pager_w, GCForeground, &gcv);
-	}
-
-	/* Hilight GC, used for monitors and labels backgrounds. */
-	gcv.foreground = (Pdepth < 2) ? style->bg : style->hi_bg;
-	style->hi_bg_gc = fvwmlib_XCreateGC(
-		dpy, Scr.pager_w, GCForeground, &gcv);
-
-	/* create the hilight desk title drawing GC */
-	gcv.foreground = (Pdepth < 2) ? style->fg : style->hi_fg;
-	if (Ffont && Ffont->font)
-		style->hi_fg_gc = fvwmlib_XCreateGC(
-			dpy, Scr.pager_w, GCForeground | GCFont, &gcv);
-	else
-		style->hi_fg_gc = fvwmlib_XCreateGC(
-			dpy, Scr.pager_w, GCForeground, &gcv);
-
-	/* create the virtual page boundary GC */
-	gcv.foreground = style->fg;
-	gcv.line_width = 1;
-	gcv.line_style = (use_dashed_separators) ? LineOnOffDash : LineSolid;
-	style->dashed_gc = fvwmlib_XCreateGC(dpy,
-		Scr.pager_w, GCForeground | GCLineStyle | GCLineWidth, &gcv);
-	if (use_dashed_separators) {
-		/* Although this should already be the default for a freshly
-		 * created GC, some X servers do not draw properly dashed
-		 * lines if the dash style is not set explicitly.
-		 */
-		dash_list[0] = 4;
-		dash_list[1] = 4;
-		XSetDashes(dpy, style->dashed_gc, 0, dash_list, 2);
-	}
-
-	/* Window borders. */
-	gcv.foreground = (style->win_cs) ?
-			 Colorset[style->win_cs].hilite : style->fg;
-	style->win_hi_gc = fvwmlib_XCreateGC(
-		dpy, Scr.pager_w, GCForeground, &gcv);
-	gcv.foreground = (style->win_cs) ?
-			 Colorset[style->win_cs].shadow : style->fg;
-	style->win_sh_gc = fvwmlib_XCreateGC(
-		dpy, Scr.pager_w, GCForeground, &gcv);
-
-	/* Focus window borders. */
-	gcv.foreground = (style->focus_cs) ?
-			 Colorset[style->focus_cs].hilite : style->fg;
-	style->focus_hi_gc = fvwmlib_XCreateGC(
-		dpy, Scr.pager_w, GCForeground, &gcv);
-	gcv.foreground = (style->focus_cs) ?
-			 Colorset[style->focus_cs].shadow : style->fg;
-	style->focus_sh_gc = fvwmlib_XCreateGC(
-		dpy, Scr.pager_w, GCForeground, &gcv);
-}
-
-void initialize_desk_windows(int desk)
-{
-	unsigned long valuemask;
-	XSetWindowAttributes attributes;
-	DeskStyle *style = Desks[desk].style;
-
-	/* Common settings for both title and desk window. */
-	attributes.colormap = Pcmap;
-	attributes.background_pixmap = default_pixmap;
-	attributes.border_pixel = style->fg;
-
-	/* Note, pixmaps are set to None to be updated later. */
-	valuemask = (CWBorderPixel | CWColormap | CWEventMask);
-	if (style->use_label_pixmap && style->cs >= 0) {
-		if (Colorset[style->cs].pixmap) {
-			valuemask |= CWBackPixmap;
-			attributes.background_pixmap = None;
-		} else {
-			valuemask |= CWBackPixel;
-			attributes.background_pixel = style->bg;
-		}
-	} else if (style->use_label_pixmap && style->bgPixmap) {
-		valuemask |= CWBackPixmap;
-		attributes.background_pixmap = None;
-	} else {
-		valuemask |= CWBackPixel;
-		attributes.background_pixel = style->bg;
-	}
-
-	attributes.event_mask = (ExposureMask | ButtonReleaseMask);
-	Desks[desk].title_w = XCreateWindow(dpy, Scr.pager_w,
-			-1, -1, desk_w, desk_h + label_h, 1,
-			CopyFromParent, InputOutput, CopyFromParent,
-			valuemask, &attributes);
-
-	/* Create desk windows. */
-	valuemask &= ~(CWBackPixel | CWBackPixmap);
-	if (style->cs >= 0) {
-		valuemask |= CWBackPixmap;
-		if (style->use_label_pixmap) {
-			attributes.background_pixmap = ParentRelative;
-			attributes.background_pixel = None;
-		} else {
-			valuemask |= CWBackPixel;
-			attributes.background_pixmap = None;
-			attributes.background_pixel = style->bg;
-		}
-	} else if (Desks[desk].style->bgPixmap) {
-		valuemask |= CWBackPixmap;
-		if (Desks[desk].style->use_label_pixmap)
-			attributes.background_pixmap = ParentRelative;
-		else
-			attributes.background_pixmap = None;
-	} else {
-		valuemask |= CWBackPixel;
-		attributes.background_pixel = Desks[desk].style->bg;
-	}
-
-	attributes.event_mask = (ExposureMask | ButtonReleaseMask |
-				 ButtonPressMask | ButtonMotionMask);
-	Desks[desk].w = XCreateWindow(dpy, Desks[desk].title_w,
-			-1, LabelsBelow ? -1 : label_h - 1, desk_w, desk_h, 1,
-			CopyFromParent, InputOutput, CopyFromParent,
-			valuemask, &attributes);
-
-	XMapRaised(dpy, Desks[desk].title_w);
-	XMapRaised(dpy, Desks[desk].w);
-}
-
-void initialize_pager(void)
-{
-  XWMHints wmhints;
-  XClassHint class1;
-  XTextProperty name;
-  unsigned long valuemask;
-  XSetWindowAttributes attributes;
-  int i = 0;
-  struct fpmonitor *fp = fpmonitor_this(NULL);
-  struct fpmonitor *m;
-  DeskStyle *style;
-
-  int VxPages = fp->virtual_scr.VxPages, VyPages = fp->virtual_scr.VyPages;
-
-  /* Set window size if not fully set by user to match */
-  /* aspect ratio of monitor(s) being shown. */
-  if ( pwindow.width == 0 || pwindow.height == 0 ) {
-	  int vWidth  = fpmonitor_get_all_widths();
-	  int vHeight = fpmonitor_get_all_heights();
-	  if (monitor_to_track != NULL || IsShared) {
-		  vWidth = fp->m->si->w;
-		  vHeight = fp->m->si->h;
-	  }
-
-	  if (pwindow.width > 0) {
-		  pwindow.height = (pwindow.width * vHeight) / vWidth +
-			  label_h * Rows + Rows;
-	  } else if (pwindow.height > label_h * Rows) {
-		  pwindow.width = ((pwindow.height - label_h * Rows + Rows) *
-				  vWidth) / vHeight + Columns;
-	  } else {
-		  pwindow.width = (VxPages * vWidth * Columns) / Scr.VScale +
-			  Columns;
-		  pwindow.height = (VyPages * vHeight * Rows) / Scr.VScale +
-			  label_h * Rows + Rows;
-	  }
-  }
-  set_desk_size(false);
-
-  if (is_transient)
-  {
-    rectangle screen_g;
-    fscreen_scr_arg fscr;
-
-    fscr.xypos.x = pwindow.x;
-    fscr.xypos.y = pwindow.y;
-    FScreenGetScrRect(
-      &fscr, FSCREEN_XYPOS,
-      &screen_g.x, &screen_g.y, &screen_g.width, &screen_g.height);
-    /* FIXME: Recalculate what to do if window is off screen. */
-    /* Leaving alone for now */
-    if (pwindow.width + pwindow.x > screen_g.x + screen_g.width)
-    {
-      pwindow.x = screen_g.x + screen_g.width - fp->m->si->w; //fp->m->virtual_scr.MyDisplayWidth;
-      xneg = true;
-    }
-    if (pwindow.height + pwindow.y > screen_g.y + screen_g.height)
-    {
-      pwindow.y = screen_g.y + screen_g.height - fp->m->si->h; //fp->m->virtual_scr.MyDisplayHeight;
-      yneg = true;
-    }
-  }
-  if (xneg)
-  {
-    sizehints.win_gravity = NorthEastGravity;
-    pwindow.x = fpmonitor_get_all_widths() - pwindow.width + pwindow.x;
-  }
-  if (yneg)
-  {
-    pwindow.y = fpmonitor_get_all_heights() - pwindow.height + pwindow.y;
-    if(sizehints.win_gravity == NorthEastGravity)
-      sizehints.win_gravity = SouthEastGravity;
-    else
-      sizehints.win_gravity = SouthWestGravity;
-  }
-
-  if(usposition)
-    sizehints.flags |= USPosition;
-
-  valuemask = (CWBackPixmap | CWBorderPixel | CWColormap | CWEventMask);
-  attributes.background_pixmap = default_pixmap;
-  attributes.border_pixel = 0;
-  attributes.colormap = Pcmap;
-  attributes.event_mask = (StructureNotifyMask);
-
-  /* destroy the temp window first, don't worry if it's the Root */
-  if (Scr.pager_w != Scr.Root)
-    XDestroyWindow(dpy, Scr.pager_w);
-  Scr.pager_w = XCreateWindow (dpy, Scr.Root, pwindow.x, pwindow.y, pwindow.width,
-			       pwindow.height, 0, Pdepth, InputOutput, Pvisual,
-			       valuemask, &attributes);
-  XSetWMProtocols(dpy,Scr.pager_w,&wm_del_win,1);
-  /* hack to prevent mapping on wrong screen with StartsOnScreen */
-  FScreenMangleScreenIntoUSPosHints(FSCREEN_XYPOS, &sizehints);
-  XSetWMNormalHints(dpy,Scr.pager_w,&sizehints);
-  if (is_transient)
-  {
-    XSetTransientForHint(dpy, Scr.pager_w, Scr.Root);
-  }
-
-  if (FlocaleTextListToTextProperty(
-	 dpy, &(Desks[0].style->label), 1, XStdICCTextStyle, &name) == 0)
-  {
-    fprintf(stderr,"%s: fatal error: cannot allocate desk name", MyName);
-    exit(0);
-  }
-
-  attributes.event_mask = (StructureNotifyMask| ExposureMask);
-  if(icon.width < 1)
-    icon.width = (pwindow.width - Columns + 1) / Columns;
-  if(icon.height < 1)
-    icon.height = (pwindow.height - Rows * label_h - Rows + 1) / Rows;
-
-  icon.width = (icon.width / (VxPages + 1)) * (VxPages + 1) + VxPages;
-  icon.height = (icon.height / (VyPages + 1)) * (VyPages + 1) + VyPages;
-  Scr.icon_w = XCreateWindow (dpy, Scr.Root, pwindow.x, pwindow.y,
-			    icon.width, icon.height,
-			    0, Pdepth, InputOutput, Pvisual, valuemask,
-			    &attributes);
-  XGrabButton(dpy, 1, AnyModifier, Scr.icon_w,
-	      True, ButtonPressMask | ButtonReleaseMask|ButtonMotionMask,
-	      GrabModeAsync, GrabModeAsync, None,
-	      None);
-  XGrabButton(dpy, 2, AnyModifier, Scr.icon_w,
-	      True, ButtonPressMask | ButtonReleaseMask|ButtonMotionMask,
-	      GrabModeAsync, GrabModeAsync, None,
-	      None);
-  XGrabButton(dpy, 3, AnyModifier, Scr.icon_w,
-	      True, ButtonPressMask | ButtonReleaseMask|ButtonMotionMask,
-	      GrabModeAsync, GrabModeAsync, None,
-	      None);
-
-  if (StartIconic) {
-	  Balloon.show = Balloon.show_in_icon;
-	  wmhints.initial_state = IconicState;
-  } else {
-	  Balloon.show = Balloon.show_in_pager;
-	  wmhints.initial_state = NormalState;
-  }
-  wmhints.flags = 0;
-  if (icon.x != -10000)
-  {
-    if (icon_xneg)
-      icon.x = fpmonitor_get_all_widths() + icon.x - icon.width;
-    if (icon.y != -10000)
-    {
-      if (icon_yneg)
-	icon.y = fpmonitor_get_all_heights() + icon.y - icon.height;
-    }
-    else
-    {
-      icon.y = 0;
-    }
-    icon_xneg = false;
-    icon_yneg = false;
-    wmhints.icon_x = icon.x;
-    wmhints.icon_y = icon.y;
-    wmhints.flags = IconPositionHint;
-  }
-  wmhints.icon_window = Scr.icon_w;
-  wmhints.input = True;
-  wmhints.flags |= InputHint | StateHint | IconWindowHint;
-
-  class1.res_name = MyName;
-  class1.res_class = "FvwmPager";
-
-  XSetWMProperties(dpy,Scr.pager_w,&name,&name,NULL,0,
-		   &sizehints,&wmhints,&class1);
-  XFree((char *)name.value);
-
-  /* change colour/font for labelling mini-windows */
-  XSetForeground(dpy, Scr.NormalGC, GetSimpleColor("black"));
-
-  if (FwindowFont != NULL && FwindowFont->font != NULL)
-    XSetFont(dpy, Scr.NormalGC, FwindowFont->font->fid);
-
-  /* Initialize DeskStyle GCs */
-  for (i = 0; i < ndesks; i++) {
-	  /* Create any missing DeskStyles. */
-	  style = FindDeskStyle(i);
-  }
-  TAILQ_FOREACH(style, &desk_style_q, entry) {
-	  initialize_desk_style_gcs(style);
-  }
-
-  for(i=0;i<ndesks;i++)
-  {
-    initialize_desk_windows(i);
-  }
-
-  TAILQ_FOREACH(m, &fp_monitor_q, entry) {
-    initialize_fpmonitor_windows(m);
-  }
-  initialize_balloon_window();
-  XMapRaised(dpy,Scr.pager_w);
-}
-
 
 void UpdateWindowShape(void)
 {
@@ -1724,42 +840,6 @@ void update_pr_transparent_subwindows(int i)
 	}
 }
 #endif
-
-/* update all the parental relative windows: pr background change */
-void update_pr_transparent_windows(void)
-{
-	int cset;
-	PagerWindow *t;
-
-	for(cset = 0; cset < ndesks; cset++) {
-		if (CSET_IS_TRANSPARENT_PR(Desks[cset].style->cs))
-			update_desk_background(cset);
-		if (CSET_IS_TRANSPARENT_PR(Desks[cset].style->hi_cs))
-			update_monitor_backgrounds(cset);
-	}
-
-	/* subordinate windows with a pr parent desk */
-	t = Start;
-	for(t = Start; t != NULL; t = t->next)
-	{
-		DeskStyle *style = FindDeskStyle(t->desk);
-
-		cset = (t == FocusWin) ? style->focus_cs : style->win_cs;
-		if (!CSET_IS_TRANSPARENT_PR(cset) ||
-		    (fAlwaysCurrentDesk &&
-		     !CSET_IS_TRANSPARENT_PR(Desks[0].style->cs)) ||
-		    (!fAlwaysCurrentDesk &&
-		     !CSET_IS_TRANSPARENT_PR(Desks[t->desk].style->cs)))
-		{
-			continue;
-		}
-		update_window_background(t);
-	}
-
-	/* ballon */
-	if (Balloon.is_mapped && CSET_IS_TRANSPARENT_PR(Balloon.cs))
-			XClearArea(dpy, Scr.balloon_w, 0, 0, 0, 0, True);
-}
 
 void MovePage()
 {
@@ -2690,243 +1770,6 @@ void MoveWindow(XEvent *Event)
 done_moving:
 	if (is_transient)
 		ExitPager(); /* does not return */
-}
-
-
-/*
- *
- *  Procedure:
- *	FvwmErrorHandler - displays info on internal errors
- *
- */
-int FvwmErrorHandler(Display *dpy, XErrorEvent *event)
-{
-#if 1
-  extern bool error_occured;
-  error_occured = true;
-  return 0;
-#else
-  /* really should just exit here... */
-  /* domivogt (07-mar-1999): No, not really. See comment above. */
-  PrintXErrorAndCoredump(dpy, event, MyName);
-  return 0;
-#endif /* 1 */
-}
-
-static void draw_window_border(PagerWindow *t, Window w, rectangle r)
-{
-	if (t == NULL)
-		return;
-
-	DeskStyle *style = FindDeskStyle(t->desk);
-
-	if (WindowBorders3d && t == FocusWin && style->focus_cs > 0) {
-		RelieveRectangle(
-				dpy, w, 0, 0, r.width - 1, r.height - 1,
-				style->focus_hi_gc, style->focus_sh_gc,
-				WindowBorderWidth);
-		return;
-	} else if (WindowBorders3d && t != FocusWin && style->win_cs > 0) {
-		RelieveRectangle(
-				dpy, w, 0, 0, r.width - 1, r.height - 1,
-				style->win_hi_gc, style->win_sh_gc,
-				WindowBorderWidth);
-		return;
-	}
-
-	if (t == FocusWin) {
-		if (style->focus_fg)
-			XSetForeground(dpy,
-				style->focus_hi_gc, style->focus_fg);
-		else
-			XSetForeground(dpy,
-				style->focus_hi_gc, Scr.focus_win_fg);
-		RelieveRectangle(
-				dpy, w, 0, 0, r.width - 1, r.height - 1,
-				style->focus_hi_gc, style->focus_hi_gc,
-				WindowBorderWidth);
-	} else {
-		if (style->win_fg)
-			XSetForeground(dpy,
-				style->win_hi_gc, style->win_fg);
-		else
-			XSetForeground(dpy,
-				style->win_hi_gc, t->text);
-		RelieveRectangle(
-				dpy, w, 0, 0, r.width - 1, r.height - 1,
-				style->win_hi_gc, style->win_hi_gc,
-				WindowBorderWidth);
-	}
-}
-
-/* draw the window label with simple greedy wrapping */
-static void label_window_wrap(PagerWindow *t, rectangle win_r)
-{
-  char *cur, *next, *end;
-  int space_width, cur_width;
-
-  space_width = FlocaleTextWidth(FwindowFont, " ", 1);
-  cur_width   = 0;
-
-  cur = next = t->window_label;
-  end = cur + strlen(cur);
-
-  while (*cur) {
-    while (*next) {
-      int width;
-      char *p;
-
-      if (!(p = strchr(next, ' ')))
-        p = end;
-
-      width = FlocaleTextWidth(FwindowFont, next, p - next );
-      if (width > win_r.width - cur_width - space_width - 2*label_border)
-        break;
-      cur_width += width + space_width;
-      next = *p ? p + 1 : p;
-    }
-
-    if (cur == next) {
-      /* word too large for window */
-      while (*next) {
-        int len, width;
-
-        len   = FlocaleStringNumberOfBytes(FwindowFont, next);
-        width = FlocaleTextWidth(FwindowFont, next, len);
-
-        if (width > win_r.width - cur_width - 2*label_border && cur != next)
-          break;
-
-        next      += len;
-        cur_width += width;
-      }
-    }
-
-    FwinString->str = fxmalloc(next - cur + 1);
-    strncpy(FwinString->str, cur, next - cur);
-    FwinString->str[next - cur] = 0;
-
-    FlocaleDrawString(dpy, FwindowFont, FwinString, 0);
-
-    free(FwinString->str);
-    FwinString->str = NULL;
-
-    FwinString->y += FwindowFont->height;
-    cur = next;
-    cur_width = 0;
-  }
-}
-
-static void do_label_window(PagerWindow *t, Window w, rectangle r)
-{
-	if (t == NULL || w == None || FwindowFont == NULL ||
-	    (MiniIcons && t->mini_icon.picture) || /* Draw image later. */
-	    t->icon_name == NULL)
-		return;
-
-	int cs;
-	Pixel pix;
-	DeskStyle *style = FindDeskStyle(t->desk);
-
-	if (t == FocusWin) {
-		pix = (style->focus_fg) ? style->focus_fg : Scr.focus_win_fg;
-		cs = style->focus_cs;
-	} else {
-		pix = (style->win_fg) ? style->win_fg : t->text;
-		cs = style->win_cs;
-	}
-	XSetForeground(dpy, Scr.NormalGC, pix);
-
-	free(t->window_label);
-	t->window_label = get_label(t, WindowLabelFormat);
-
-	if (FftSupport)
-		XClearWindow(dpy, w);
-	FwinString->win = w;
-	FwinString->gc = Scr.NormalGC;
-	FwinString->flags.has_colorset = False;
-	if (cs >= 0) {
-		FwinString->colorset = &Colorset[cs];
-		FwinString->flags.has_colorset = True;
-	}
-	FwinString->x = label_border;
-	FwinString->y = FwindowFont->ascent+2;
-
-	label_window_wrap(t, r);
-}
-
-static void do_picture_window(PagerWindow *t, Window w, rectangle r)
-{
-	if (t == NULL || !MiniIcons || !(t->mini_icon.picture))
-		return;
-
-	int iconX, iconY;
-	int src_x = 0, src_y = 0;
-	int dest_w, dest_h;
-	int cs;
-	FvwmRenderAttributes fra;
-	DeskStyle *style = FindDeskStyle(t->desk);
-
-	cs = (t != FocusWin) ? style->win_cs : style->focus_cs;
-	dest_w = t->mini_icon.width;
-	dest_h = t->mini_icon.height;
-	if (r.width > t->mini_icon.width) {
-		iconX = (r.width - t->mini_icon.width) / 2;
-	} else if (r.width < t->mini_icon.width) {
-		iconX = 0;
-		src_x = (t->mini_icon.width - r.width) / 2;
-		dest_w = r.width;
-	} else {
-		iconX = 0;
-	}
-	if (r.height > t->mini_icon.height) {
-		iconY = (r.height - t->mini_icon.height) / 2;
-	} else if (r.height < t->mini_icon.height) {
-		iconY = 0;
-		src_y = (t->mini_icon.height - r.height) / 2;
-		dest_h = r.height;
-	} else {
-		iconY = 0;
-	}
-
-	fra.mask = FRAM_DEST_IS_A_WINDOW;
-	if (t->mini_icon.alpha != None ||
-	    (cs >= 0 && Colorset[cs].icon_alpha_percent < 100))
-	{
-		XClearArea(dpy, w, iconX, iconY, t->mini_icon.width,
-			   t->mini_icon.height, False);
-	}
-
-	if (cs >= 0) {
-		fra.mask |= FRAM_HAVE_ICON_CSET;
-		fra.colorset = &Colorset[cs];
-	}
-
-	PGraphicsRenderPicture(dpy, w, &t->mini_icon, &fra, w,
-			       Scr.MiniIconGC, None, None, src_x, src_y,
-			       t->mini_icon.width - src_x,
-			       t->mini_icon.height - src_y,
-			       iconX, iconY, dest_w, dest_h, False);
-}
-
-void update_pager_window_decor(PagerWindow *t)
-{
-	do_label_window(t, t->PagerView, t->pager_view);
-	do_picture_window(t, t->PagerView, t->pager_view);
-	draw_window_border(t, t->PagerView, t->pager_view);
-}
-
-void update_icon_window_decor(PagerWindow *t)
-{
-	do_label_window(t, t->IconView, t->icon_view);
-	do_picture_window(t, t->IconView, t->icon_view);
-	draw_window_border(t, t->IconView, t->icon_view);
-}
-
-void update_window_decor(PagerWindow *t)
-{
-	update_pager_window_decor(t);
-	update_icon_window_decor(t);
 }
 
 void IconMoveWindow(XEvent *Event, PagerWindow *t)
